// apps/client/src/utils/modalKeys.ts
export enum ModalKeys {
    Login = "login",
    Register = "register",
    LoginConfirm = "loginConfirm",
    RegisterConfirm = "registerConfirm",
    AddWorkspace = "addWorkspace",
    AddWorkspaceConfirm = "addWorkspaceConfirm",
    AddColumn = "addColumn",
    AddColumnConfirm = "addColumnConfirm",
    AddTask = "addTask",
    AddTaskConfirm = "addTaskConfirm",
}